css/image-fit/js/inherited/008.html

<!doctype html>
<title>'inherit'ing from &lt;video></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video style=-o-object-fit:auto></video>
<script>
test(function() {
  var e = document.body.firstChild;
  e.appendChild(document.createElement('div'));
  e.firstChild.style.OObjectFit = 'inherit';
  e.firstChild.style.OObjectPosition = 'inherit';
  assert_equals(getComputedStyle(e.firstChild, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e.firstChild, '').OObjectPosition, '50% 50%');
});
</script>
